+++ utilities/getHash.js
@@ -1,7 +1,6 @@
 export default function getHash(agent) {
   if (agent) {
     return agent.split("id/").pop()
-  } else {
-    throw new Error("No agent provided")
   }
+  throw new Error("No agent provided")
 }
