+++ utilities/isDefaultRole.js
@@ -1,26 +1,19 @@
 import Group from "../classes/Group/Group.js"
 
 export default function scrubDefaultRoles(roleName) {
-  
   const defaultRoles = Object.keys(Group.defaultRoles)
-  if(Array.isArray(roleName)) {
+  if (Array.isArray(roleName)) {
     roleName = roleName.filter(roleString => {
-      if(typeof roleString !== "string") {
-        throw new Error("Expecting a RolesMap and not an Array.")
-      }
+      if (typeof roleString !== "string") throw new Error("Expecting a RolesMap and not an Array.")
       return !defaultRoles.includes(roleString)
     })
     return roleName.length !== 0 ? roleName : false
   }
-
   if (typeof roleName === "object") {
     for (const role of Object.keys(roleName)) {
-      if (defaultRoles.includes(role)) {
-        delete roleName[role]
-      }
+      if (defaultRoles.includes(role)) delete roleName[role]
     }
     return Object.keys(roleName).length !== 0 ? roleName : false
   }
-
   return false
 }
