+++ utilities/validateEmail.js
@@ -1,6 +1,4 @@
-const validEmail = new RegExp(
-  /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/
-)
+const validEmail = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/
 
 export function isValidEmail(email) {
   return validEmail.test(email)
